% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopairtest-package.R
\docType{package}
\name{autopairtest-package}
\alias{autopairtest}
\alias{autopairtest-package}
\title{autopairtest: Automatic Pairwise Statistical Testing for Continuous and Categorical Outcomes}
\description{
Provides two functions for automatic, unpaired pairwise comparisons of independent groups. 'pairwise_auto_test()' compares continuous outcomes between all pairs of groups, automatically selecting Welch t-test (if both groups pass Shapiro-Wilk normality) or Wilcoxon rank-sum test. 'pairwise_auto_cat_test()' compares categorical outcomes, automatically selecting chi-square or Fisher's exact test based on expected cell counts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://uhanwu.github.io/autopairtest/}
}

}
\author{
\strong{Maintainer}: Yuhan (Max) Wu \email{wuyuhanm@gmail.com}

Authors:
\itemize{
  \item Yuhan (Max) Wu \email{wuyuhanm@gmail.com}
}

}
\keyword{internal}
